[network] Move discovery network into discovery
To make networking as effective as possible, it makes sense to host the
interface into networking in the code itself. This allows for binding
the types sent across the network into local types. While this isn't
done in this PR, this is part of the future goals. This PR is just a
bridge to that future.
